state-management/conversation-history.py
import logging 
from strands import Agent

logging.getLogger("strands").setLevel(logging.DEBUG)
logging.basicConfig(
    format="%(levelname)s | %(name)s | %(message)s",
    handlers=[logging.StreamHandler()],
)

agent = Agent()

agent("Hello")

# Access the conversation history
print(agent.messages)